feat(cosmwasm): continue spec compliance
* Add `add_owners` and `set_owners` method to add and remove
owners to the interpreter.

* Change `AssetId` to u128 for compliance with Picasso.

* Add `relayer_address` to `InstantiateMsg`.

* Fix `relayer_address` binding and instead of returning
`user_id` in `BindingValue::Relayer`, return the relayer
address.

+++ code/xcvm/cosmwasm/contracts/asset-registry/src/state.rs
@@ -1,6 +1,6 @@
 use cosmwasm_std::Addr;
 use cw_storage_plus::Map;
 
-pub type XcvmAssetId = u64;
+pub type XcvmAssetId = u128;
 
 pub const ASSETS: Map<XcvmAssetId, Addr> = Map::new("assets");
